Rename a directory
src/arrow/sparse_tensor -> src/arrow/tensor
