+++ cpp/src/arrow/CMakeLists.txt
@@ -144,14 +144,14 @@ set(ARROW_SRCS
     result.cc
     scalar.cc
     sparse_tensor.cc
-    sparse_tensor/coo_converter.cc
-    sparse_tensor/csc_converter.cc
-    sparse_tensor/csf_converter.cc
-    sparse_tensor/csr_converter.cc
     status.cc
     table.cc
     table_builder.cc
     tensor.cc
+    tensor/coo_converter.cc
+    tensor/csc_converter.cc
+    tensor/csf_converter.cc
+    tensor/csr_converter.cc
     type.cc
     visitor.cc
     c/bridge.cc
