+++ cpp/src/arrow/tensor/coo_converter.cc
@@ -15,24 +15,39 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "arrow/sparse_tensor/converter.h"
+#include "arrow/tensor/converter.h"
 
+#include "arrow/buffer_builder.h"
 #include "arrow/visitor_inline.h"
 
 namespace arrow {
 namespace internal {
 namespace {
 
+inline void IncrementIndex(std::vector<int64_t>& coord,
+                           const std::vector<int64_t>& shape) {
+  const int64_t ndim = shape.size();
+  ++coord[ndim - 1];
+  if (coord[ndim - 1] == shape[ndim - 1]) {
+    int64_t d = ndim - 1;
+    while (d > 0 && coord[d] == shape[d]) {
+      coord[d] = 0;
+      ++coord[d - 1];
+      --d;
+    }
+  }
+}
+
 // ----------------------------------------------------------------------
-// SparseTensorConverter for SparseCSCIndex
+// SparseTensorConverter for SparseCOOIndex
 
 template <typename TYPE>
-class SparseCSCMatrixConverter {
+class SparseCOOTensorConverter {
  public:
   using NumericTensorType = NumericTensor<TYPE>;
   using value_type = typename NumericTensorType::value_type;
 
-  SparseCSCMatrixConverter(const NumericTensorType& tensor,
+  SparseCOOTensorConverter(const NumericTensorType& tensor,
                            const std::shared_ptr<DataType>& index_value_type,
                            MemoryPool* pool)
       : tensor_(tensor), index_value_type_(index_value_type), pool_(pool) {}
@@ -40,23 +55,17 @@ class SparseCSCMatrixConverter {
   template <typename IndexValueType>
   Status Convert() {
     using c_index_value_type = typename IndexValueType::c_type;
-    RETURN_NOT_OK(CheckMaximumValue(std::numeric_limits<c_index_value_type>::max()));
     const int64_t indices_elsize = sizeof(c_index_value_type);
 
     const int64_t ndim = tensor_.ndim();
-    if (ndim > 2) {
-      // LCOV_EXCL_START: The following invalid causes program failure.
-      return Status::Invalid("Invalid tensor dimension");
-      // LCOV_EXCL_STOP
-    }
-
-    const int64_t nr = tensor_.shape()[0];
-    const int64_t nc = tensor_.shape()[1];
     int64_t nonzero_count = -1;
     RETURN_NOT_OK(tensor_.CountNonZero(&nonzero_count));
 
-    std::shared_ptr<Buffer> indptr_buffer;
     std::shared_ptr<Buffer> indices_buffer;
+    RETURN_NOT_OK(
+        AllocateBuffer(pool_, indices_elsize * ndim * nonzero_count, &indices_buffer));
+    c_index_value_type* indices =
+        reinterpret_cast<c_index_value_type*>(indices_buffer->mutable_data());
 
     std::shared_ptr<Buffer> values_buffer;
     RETURN_NOT_OK(
@@ -64,40 +73,36 @@ class SparseCSCMatrixConverter {
     value_type* values = reinterpret_cast<value_type*>(values_buffer->mutable_data());
 
     if (ndim <= 1) {
-      return Status::NotImplemented("TODO for ndim <= 1");
+      const value_type* data = reinterpret_cast<const value_type*>(tensor_.raw_data());
+      const int64_t count = ndim == 0 ? 1 : tensor_.shape()[0];
+      for (int64_t i = 0; i < count; ++i, ++data) {
+        if (*data != 0) {
+          *indices++ = static_cast<c_index_value_type>(i);
+          *values++ = *data;
+        }
+      }
     } else {
-      RETURN_NOT_OK(AllocateBuffer(pool_, indices_elsize * (nc + 1), &indptr_buffer));
-      auto* indptr = reinterpret_cast<c_index_value_type*>(indptr_buffer->mutable_data());
+      const std::vector<int64_t>& shape = tensor_.shape();
+      std::vector<int64_t> coord(ndim, 0);  // The current logical coordinates
 
-      RETURN_NOT_OK(
-          AllocateBuffer(pool_, indices_elsize * nonzero_count, &indices_buffer));
-      auto* indices =
-          reinterpret_cast<c_index_value_type*>(indices_buffer->mutable_data());
-
-      c_index_value_type k = 0;
-      *indptr++ = 0;
-      for (int64_t j = 0; j < nc; ++j) {
-        for (int64_t i = 0; i < nr; ++i) {
-          const value_type x = tensor_.Value({i, j});
-          if (x != 0) {
-            *values++ = x;
-            *indices++ = static_cast<c_index_value_type>(i);
-            k++;
+      for (int64_t n = tensor_.size(); n > 0; n--) {
+        const value_type x = tensor_.Value(coord);
+        if (tensor_.Value(coord) != 0) {
+          *values++ = x;
+          // Write indices in row-major order.
+          for (int64_t i = 0; i < ndim; ++i) {
+            *indices++ = static_cast<c_index_value_type>(coord[i]);
           }
         }
-        *indptr++ = k;
+        IncrementIndex(coord, shape);
       }
     }
 
-    std::vector<int64_t> indptr_shape({nc + 1});
-    std::shared_ptr<Tensor> indptr_tensor =
-        std::make_shared<Tensor>(index_value_type_, indptr_buffer, indptr_shape);
-
-    std::vector<int64_t> indices_shape({nonzero_count});
-    std::shared_ptr<Tensor> indices_tensor =
-        std::make_shared<Tensor>(index_value_type_, indices_buffer, indices_shape);
-
-    sparse_index = std::make_shared<SparseCSCIndex>(indptr_tensor, indices_tensor);
+    // make results
+    const std::vector<int64_t> indices_shape = {nonzero_count, ndim};
+    const std::vector<int64_t> indices_strides = {indices_elsize * ndim, indices_elsize};
+    sparse_index = std::make_shared<SparseCOOIndex>(std::make_shared<Tensor>(
+        index_value_type_, indices_buffer, indices_shape, indices_strides));
     data = values_buffer;
 
     return Status::OK();
@@ -119,54 +124,40 @@ class SparseCSCMatrixConverter {
 
 #undef CALL_TYPE_SPECIFIC_CONVERT
 
-  std::shared_ptr<SparseCSCIndex> sparse_index;
+  std::shared_ptr<SparseCOOIndex> sparse_index;
   std::shared_ptr<Buffer> data;
 
  private:
   const NumericTensorType& tensor_;
   const std::shared_ptr<DataType>& index_value_type_;
   MemoryPool* pool_;
-
-  template <typename c_value_type>
-  inline Status CheckMaximumValue(const c_value_type type_max) const {
-    if (static_cast<int64_t>(type_max) < tensor_.shape()[1]) {
-      // LCOV_EXCL_START: The following invalid causes program failure.
-      return Status::Invalid("The bit width of the index value type is too small");
-      // LCOV_EXCL_STOP
-    }
-    return Status::OK();
-  }
-
-  inline Status CheckMaximumValue(const int64_t) const { return Status::OK(); }
-
-  inline Status CheckMaximumValue(const uint64_t) const { return Status::OK(); }
 };
 
 // ----------------------------------------------------------------------
 // Instantiate templates
 
-template class SparseCSCMatrixConverter<UInt8Type>;
-template class SparseCSCMatrixConverter<UInt16Type>;
-template class SparseCSCMatrixConverter<UInt32Type>;
-template class SparseCSCMatrixConverter<UInt64Type>;
-template class SparseCSCMatrixConverter<Int8Type>;
-template class SparseCSCMatrixConverter<Int16Type>;
-template class SparseCSCMatrixConverter<Int32Type>;
-template class SparseCSCMatrixConverter<Int64Type>;
-template class SparseCSCMatrixConverter<HalfFloatType>;
-template class SparseCSCMatrixConverter<FloatType>;
-template class SparseCSCMatrixConverter<DoubleType>;
+template class SparseCOOTensorConverter<UInt8Type>;
+template class SparseCOOTensorConverter<UInt16Type>;
+template class SparseCOOTensorConverter<UInt32Type>;
+template class SparseCOOTensorConverter<UInt64Type>;
+template class SparseCOOTensorConverter<Int8Type>;
+template class SparseCOOTensorConverter<Int16Type>;
+template class SparseCOOTensorConverter<Int32Type>;
+template class SparseCOOTensorConverter<Int64Type>;
+template class SparseCOOTensorConverter<HalfFloatType>;
+template class SparseCOOTensorConverter<FloatType>;
+template class SparseCOOTensorConverter<DoubleType>;
 
 // ----------------------------------------------------------------------
 
 template <typename TYPE>
-Status MakeSparseCSCMatrixFromTensor(const Tensor& tensor,
+Status MakeSparseCOOTensorFromTensor(const Tensor& tensor,
                                      const std::shared_ptr<DataType>& index_value_type,
                                      MemoryPool* pool,
                                      std::shared_ptr<SparseIndex>* out_sparse_index,
                                      std::shared_ptr<Buffer>* out_data) {
   NumericTensor<TYPE> numeric_tensor(tensor.data(), tensor.shape(), tensor.strides());
-  SparseCSCMatrixConverter<TYPE> converter(numeric_tensor, index_value_type, pool);
+  SparseCOOTensorConverter<TYPE> converter(numeric_tensor, index_value_type, pool);
   RETURN_NOT_OK(converter.Convert());
 
   *out_sparse_index = checked_pointer_cast<SparseIndex>(converter.sparse_index);
@@ -176,18 +167,18 @@ Status MakeSparseCSCMatrixFromTensor(const Tensor& tensor,
 
 }  // namespace
 
-#define MAKE_SPARSE_CSC_MATRIX_FROM_TENSOR(TYPE_CLASS)      \
+#define MAKE_SPARSE_TENSOR_FROM_TENSOR(TYPE_CLASS)          \
   case TYPE_CLASS##Type::type_id:                           \
-    return MakeSparseCSCMatrixFromTensor<TYPE_CLASS##Type>( \
+    return MakeSparseCOOTensorFromTensor<TYPE_CLASS##Type>( \
         tensor, index_value_type, pool, out_sparse_index, out_data);
 
-Status MakeSparseCSCMatrixFromTensor(const Tensor& tensor,
+Status MakeSparseCOOTensorFromTensor(const Tensor& tensor,
                                      const std::shared_ptr<DataType>& index_value_type,
                                      MemoryPool* pool,
                                      std::shared_ptr<SparseIndex>* out_sparse_index,
                                      std::shared_ptr<Buffer>* out_data) {
   switch (tensor.type()->id()) {
-    ARROW_GENERATE_FOR_ALL_NUMERIC_TYPES(MAKE_SPARSE_CSC_MATRIX_FROM_TENSOR);
+    ARROW_GENERATE_FOR_ALL_NUMERIC_TYPES(MAKE_SPARSE_TENSOR_FROM_TENSOR);
       // LCOV_EXCL_START: ignore program failure
     default:
       return Status::TypeError("Unsupported Tensor value type");
@@ -195,7 +186,7 @@ Status MakeSparseCSCMatrixFromTensor(const Tensor& tensor,
   }
 }
 
-#undef MAKE_SPARSE_CSC_MATRIX_FROM_TENSOR
+#undef MAKE_SPARSE_TENSOR_FROM_TENSOR
 
 }  // namespace internal
 }  // namespace arrow
